Rename chorenzo-config.utils.ts to config.utils.ts
Remove redundant 'chorenzo' prefix from config utility filename and update all imports accordingly. This makes the filename more concise while maintaining clarity.
