+++ src/commands/recipes.ts
@@ -16,11 +16,11 @@ import {
 } from '~/types/apply';
 import { Recipe, RecipeDependency } from '~/types/recipe';
 import { WorkspaceState } from '~/types/state';
-import { chorenzoConfig } from '~/utils/chorenzo-config.utils';
 import {
   CodeChangesEventHandlers,
   executeCodeChangesOperation,
 } from '~/utils/code-changes-events.utils';
+import { chorenzoConfig } from '~/utils/config.utils';
 import { cloneRepository } from '~/utils/git-operations.utils';
 import { normalizeRepoIdentifier } from '~/utils/git.utils';
 import { readJson } from '~/utils/json.utils';
